refactor: page definitions to each page for isolation (partial)
