+++ experiments/veo-app/pages/library.py
@@ -47,6 +47,13 @@ from components.library.video_grid_item import video_grid_item
 from components.library.grid_parts import render_video_pills, render_image_pills, render_audio_pills, render_video_preview, render_image_preview, render_audio_preview
 from state.state import AppState
 
+@me.page(path="/library", title="GenMedia Creative Studio - Library")
+def library_page():
+    """Main Page."""
+    state = me.state(AppState)
+    with page_scaffold(page_name="library"):  # pylint: disable=not-context-manager
+        library_content(state)
+
 
 @me.stateclass
 @dataclass
@@ -765,6 +772,6 @@ def on_refresh_click(e: me.ClickEvent):
 def page():
     """The main entry point for the library page."""
     app_state = me.state(AppState)
-    with me.box(style=me.Style(display="flex", flex_direction="row")):
+    with page_scaffold(page_name="library"):
         library_content(app_state)
